Fix retrieval of top image opacity

+++ src/scripts/h5p-agamotto-images.js
@@ -150,7 +150,7 @@ class Images {
    * @return {number} Opacity of top image.
    */
   getTopOpacity() {
-    return parseInt(this.imageTop.style.opacity || '');
+    return parseFloat(this.imageTop.style.opacity || '');
   }
 
   /**
